+++ src/test/ExSln2/c99/gen/app_Main.c
@@ -6,7 +6,6 @@
 #include "app_Main.h"
 #include "hal_Helper.h"
 #include <stdint.h>
-#include <stdint.h>
 #include <string.h>