+++ src/test/ExSln2/c99/gen/hal_SizeOfEx.h
@@ -5,7 +5,6 @@
 #pragma once
 
 #include <stdint.h>
-#include <stdint.h>
 
 
 // Class has no fields. No struct generated.