+++ src/test/ExSln2/c99/gen/ts__CArrayMemEx.h
@@ -6,7 +6,6 @@
 
 #include "ts__CArrayMemEx_Bike.h"
 #include <stdint.h>
-#include <stdint.h>