+++ CHANGELOG.md
@@ -23,7 +23,7 @@ You can easily find changes by searching for `"### Change"` in this markdown fil
 Test/interim releases are not documented here.
 
 
-## [0.???]
+## [0.6.2]
 ### Fixed
 - fix delegate and pointer type parameters
     - https://github.com/fin-language/fin/issues/86